v1.15
Added files in anticipation for Results screen asset updates that still need to be made.
Updated the rhythm game feedback assets for when the user hits a based/mid/cringe note.
Updated the rhythm game assets for when the user is in the Peak or Danger status.
Added a notification at the end of the song in the rhythm game; when the user completes the level, they'll see feedback based on whether they completed the level, completed the level with a full combo, or completed the level with all notes perfectly hit.
Added a progress bar at the bottom of the screen in the rhythm game levels.