v1.16
Fixed an issue with NPCs not darkening properly in the VS animation at the start of the rhythm scenes.
Added animations for the health bar's heart upon hitting a note, hitting a note when the user is Peaking, and missing a note.
Added missing NPC chibis for Level 3's rhythm game scene.
Updated the threshold indicator; now animates upon hitting the required Coolness, or dropping below it.
Updated the background for certain dialogue scenes.
Added special cases for clothing overlay layering on the dress-up screen (unfinished).