v1.207
Fixed an issue with the NPCs not going from dark to light on the VS animations.
"Fixed" an issue with a particle system causing a lag spike in the rhythm game upon hitting 100 combo.
Fixed an issue where the base model for Jassmea was not showing on the Results screen.
Fixed an issue where certain NPCs on the Y2K level were not positioned correctly for both the VS animation and the associated chibis.
Adjusted the timing of the VS animation in all levels to fit better with the transitions.
Adjusted the timing that the song begins in all levels to fit better with the transitions.