v1.26
Fixed an issue with certain misleading dialogue in Day 3 after the user clears the song.
Fixed a layering issue in the results screen when using certain pants.
Added an animation to the title screen.
Adjusted the size of Jassmea and all NPCs in all rhythm game scenes.
Adjusted the combo counter in all rhythm game scenes.
Adjusted the song on the title screen.