v1.27 - Particle additions
Added a particle system in stage select for a potential difficulty selection animation effect.
Added two particle systems in dress up scene to potentially be implemented. One of the particle systems is for when a category icon button is hovered over it will show sparkles on top. Second particle system is for when players select the maximum correct clothing options to show sparkles on top of the bar that is maxed out.